Draft 3.0
Rearranged the garages so they are organized by garage 1-5

Added links to the Insight commercial website and email in the about us
section. Also, updated email link.

Changed the thumbnail images for the locations.

Added custom map markers.

Added a key for the custom map markers.

Improved the info windows that pop up when you click the location
markers.