Draft 5.0
Changed the folder structure and added .htaccess file to clip of the
index.html in the url.

Changed all of the links on the page from relative to absolute path
links.

Added available now stamps on the location boxes on in the location
section of the home page.

Changed the map marker color and added a link to the new Mahaffie street
page.

Added a page for a new location.

Made the footer the same on the location pages as it is on the homepage.

Centered the photo on the location pages landing and made it fade into
view when the page loads.

Put a shadow box around the text on the location pages and made it fade
into view when the page loads.

Added a new unit size and button on the Dennis Olathe page.