Draft 5.1
Changed which locations had the available now tag.

Changed how the location boxes respond to mobile screens. Also changed
the layout on desktops.

Changed the price on the double unit at dennis.

Put coming soon images on the mahaffie page.

Added the elevation image on the landing of the mahaffie page.