Mobile Friendly
Changed the location features boxes to fit to the size of their content
so text wouldn't overflow.

Added a picture and a new link in the garage V page to show the prices
of each unit.

Made the about us section fit to the size of its content and fixed the
issue that was causing the image that would slide in to get cut off on
some screens.

Added a cool fade to the about us section.