KCMO Location
Added a new location to the site.
Added the map marker for the new location.
Added a new location card on the home page.
Changed one of the background images on the home page.
Added the new page for the new location.
Added the details for the new location on the KCMO page.