Location Details Work
Added banner images to locations detail page.
Added a map on the location detal page.
Added unit summary information to the location schema and displaying it in the unit information section.
Added an image carousel to the location detail page.
Added contact information to the location detail page.
Updated all references to images in the location schema to use a new schema for images.