Payment Processing
Hooked up payment processing to show the payment component in a modal when user clicks the make payment button.
Got rid of the route to the payment page because now payments will be handled on individual location pages.
Divided the payment markup field into two fields one for percentage and one for fixed fee.