Location API Changes
Locations now have a flag for if they are in draft state.
The query for getting all locations to show on the map will only fetch locations that are not drafts.
Location detail page will also only show locations that are not drafts.
Locations have a createdBy field that is associated with the user record for who created the location.
Updated the endpoint for getting all locations in the portal and add endpoint for getting a single location in the portal.