CORS, Previews and Fixes
Fixed the logic for getting a location by id.
Added routes for updating locations and creating previews and viewing previews.
Added a schema for previews where records expire in 30 minutes.
Added isPublished flag for locations to enable hiding a location but not making it a draft.
Edited LocationDetail component to look for previews if it is at the preview route.
Added more logic checks to prevent errors when all location data is not present.
Enhanced the cors logic to allow for preflight requests.