Unit Availabilites On Directory
Now making it easier for users to see what locations have units that are available to rent on the directory page.
An alert is shown at the top of the page if there are units available and each location has a badge indicating that units are available there in both the list and ontheir map popover.