Video Upload Updates
Updated to the latest version of node.
Updated the client to the latest version of react with create-react-app.
Updated leaflet.
Removed node-sass in place of sass.
Using a different library for marker cluster groups that support the newer version of react.
Added react-player for displaying videos.
Added video player to the location detail page.
Added a new schema for video and added it to the location schema.
Returning the upload preset used for video files on Cloudinary from the /image-upload-credentials API endpoint.