Seminar updates & Video source change
Updated seminar dates, profile image, and wording on the schedule section.
Added a new field on the attendance form to gather if people will be attending in person or online.
Updated the automated email that gets sent out to include the new attendance type field on the form.
Changed how the video is loaded on the about us page to no longer use data bandwidth from Cloudinary.